+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MovieReservationCompleteActivity"
+            android:name=".view.MovieReservationCompleteActivity"
             android:exported="false" />
         <activity
-            android:name=".MovieDetailActivity"
+            android:name=".view.MovieDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".MovieMainActivity"
+            android:name=".view.MovieMainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
