+++ app/src/main/java/woowacourse/movie/model/MovieRepository.kt
@@ -1,4 +1,6 @@
-package woowacourse.movie
+package woowacourse.movie.model
+
+import woowacourse.movie.R
 
 class MovieRepository() {
     private var _movies =
