+++ app/src/main/java/woowacourse/movie/model/ReservationCount.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.model
 
 class ReservationCount {
     var count: Int = 1
