+++ app/src/main/res/layout/activity_movie_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieMainActivity">
+    tools:context=".view.MovieMainActivity">
 
     <ListView
         android:id="@+id/movieList"
